Implement decoding of decimal values
RFC8428 specifies that all numeric types can be a number or a decimal
fraction. The latter type is not implemented by Codec, and might lead to
a loss of precision when converted to float. Additionally, converting
the integer types to floating-point numbers loses additional precision.

Add a `Decimal` type and `Numeric` interface, and change all 'numeric'
elements in `Record` to `Numeric` so they can contain integers, floats
and the (new) `Decimal` type.